+++ ksmartteam01/src/main/java/anypet/ks44team01/Ksmartteam01Application.java
@@ -1,0 +1,13 @@
+package anypet.ks44team01;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Ksmartteam01Application {
+
+	public static void main(String[] args) {
+		SpringApplication.run(Ksmartteam01Application.class, args);
+	}
+
+}
