+++ ksmartteam01/src/main/java/anypet/ks44team01/ServletInitializer.java
@@ -1,0 +1,13 @@
+package anypet.ks44team01;
+
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+public class ServletInitializer extends SpringBootServletInitializer {
+
+	@Override
+	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+		return application.sources(Ksmartteam01Application.class);
+	}
+
+}
