+++ ksmartteam01/src/main/java/anypet/ks44team01/company/controller/CompanyMemberController.java
@@ -1,0 +1,5 @@
+package anypet.ks44team01.company.controller;
+
+public class CompanyMemberController {
+
+}
