+++ ksmartteam01/src/main/java/anypet/ks44team01/controller/MainController.java
@@ -1,0 +1,18 @@
+package anypet.ks44team01.controller;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+
+
+@Controller
+public class MainController {
+	@GetMapping("/")
+	public String main(Model model) {
+		return"main";
+	}
+	
+	@GetMapping("index")
+	public String home(Model model) {
+		return"admin/index";
+	}
+}
