+++ ksmartteam01/src/main/java/anypet/ks44team01/user/controller/AccommodationListController.java
@@ -1,0 +1,14 @@
+package anypet.ks44team01.user.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/accommodation")
+public class AccommodationListController {
+	@GetMapping("/insertReservation")
+	public String insertReservation() {
+		return "/user/accommodation/insertReservation";
+	}
+}
