+++ ksmartteam01/src/main/java/anypet/ks44team01/user/controller/UserGoodsPayController.java
@@ -1,0 +1,15 @@
+package anypet.ks44team01.user.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/user/goods")
+public class UserGoodsPayController {
+
+	@GetMapping("/goodsPay")
+	public String goodsPay() {
+		return "/user/goods/goodsPay";
+	}
+}
