+++ ksmartteam01/src/main/java/anypet/ks44team01/user/controller/UserMemberController.java
@@ -1,0 +1,5 @@
+package anypet.ks44team01.user.controller;
+
+public class UserMemberController {
+
+}
