+++ ksmartteam01/src/main/java/anypet/ks44tteam01/admin/controller/AdminMemberController.java
@@ -1,0 +1,5 @@
+package anypet.ks44tteam01.admin.controller;
+
+public class AdminMemberController {
+
+}
