--- ksmartteam01/src/main/java/anypet/ksmartteam01/Ksmartteam01Application.java
@@ -1,13 +1,0 @@
-package anypet.ksmartteam01;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Ksmartteam01Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Ksmartteam01Application.class, args);
-	}
-
-}
