--- ksmartteam01/src/main/java/anypet/ksmartteam01/ServletInitializer.java
@@ -1,13 +1,0 @@
-package anypet.ksmartteam01;
-
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-
-public class ServletInitializer extends SpringBootServletInitializer {
-
-	@Override
-	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(Ksmartteam01Application.class);
-	}
-
-}
