--- ksmartteam01/src/main/java/anypet/ksmartteam01/admin/controller/AdminBoardController.java
@@ -1,5 +1,0 @@
-package anypet.ksmartteam01.admin.controller;
-
-public class AdminBoardController {
-
-}
