--- ksmartteam01/src/main/java/anypet/ksmartteam01/company/controller/CompanyGoodsPayController.java
@@ -1,5 +1,0 @@
-package anypet.ksmartteam01.company.controller;
-
-public class CompanyGoodsPayController {
-
-}
