--- ksmartteam01/src/main/java/anypet/ksmartteam01/company/controller/CompanyMemberController.java
@@ -1,5 +1,0 @@
-package anypet.ksmartteam01.company.controller;
-
-public class CompanyMemberController {
-
-}
