--- ksmartteam01/src/main/java/anypet/ksmartteam01/user/controller/UserMemberController.java
@@ -1,5 +1,0 @@
-package anypet.ksmartteam01.user.controller;
-
-public class UserMemberController {
-
-}
