+++ ksmartteam01/src/main/resources/templates/fragments/admin/leftSide.html
@@ -6,7 +6,7 @@
 		<div class="left_col scroll-view">
 			<div class="navbar nav_title" style="border: 0;">
 				<a href="index.html" class="site_title"><i class="fa fa-paw"></i>
-					<span>Anipet</span></a>
+					<span>Anypet</span></a>
 			</div>
 
 			<div class="clearfix"></div>
